Break out `*-trace` commands
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`convert-trace` and `view-trace` commands into a `trace` subpackage.
