+++ pkg/cmd/pulumi/trace/convert.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2021, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package trace
 
 import (
 	"context"
@@ -701,7 +701,7 @@ func exportTraceToOtel(querier appdash.Queryer, ignoreLogSpans bool) error {
 	return exporter.Shutdown(context.Background())
 }
 
-func newConvertTraceCmd() *cobra.Command {
+func NewConvertTraceCmd() *cobra.Command {
 	var otel bool
 	var ignoreLogSpans bool
 	var quantum time.Duration
